Deal with multiple fg vals for same life stage
Dealt with the case where multiple functional groups were returned for the same life stage (e.g. AphiaID 123071 returns 'benthos' and 'not applicable' as adult functional groups). Now these stages are returned as 'adult' and 'adult_2' - and more generally, any replicate stages have a numeric suffix added (except the first one), in this code:
``` #' deal with cases where multiple records are returned for the same life stage:
        #' add a suffix to subsequent identical stages (adult_2, etc.)        
        life_stage <- life_stage %>% group_by(stage) %>% mutate(nth_stage_val = 1:n()) %>% 
          ungroup() %>% 
          mutate(stage = case_when(
            nth_stage_val == 1 ~ stage,
            TRUE ~ paste(stage, nth_stage_val, sep = "_")
          )
          )
```
NB this is only implemented in cases where life stage is provided. There may still be cases with no life stage but multiple conflicting functional groups, which will need to be dealt with elsewhere in the function.